requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#

-e .
astroid==2.3.3            # via pylint
atomicwrites==1.3.0       # via pytest
attrs==19.3.0             # via pytest
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
colorama==0.4.1           # via pylint, pytest
coverage==4.5.4
entrypoints==0.3          # via flake8
flake8==3.7.9
idna==2.8                 # via requests
importlib-metadata==0.23  # via pluggy, pytest
isort==4.3.21             # via pylint
lazy-object-proxy==1.4.3  # via astroid
mccabe==0.6.1             # via flake8, pylint
mock==3.0.5
more-itertools==7.2.0     # via pytest, zipp
packaging==19.2           # via pytest
pluggy==0.13.1            # via pytest
py==1.8.0                 # via pytest
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pyparsing==2.4.5          # via packaging
pytest-cov==2.8.1
pytest-runner==5.2
pytest==5.3.0
requests==2.22.0
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.5
six==1.13.0               # via astroid, mock, packaging
typed-ast==1.4.0          # via astroid
urllib3==1.25.7           # via requests
wcwidth==0.1.7            # via pytest
wrapt==1.11.2             # via astroid
zipp==0.6.0               # via importlib-metadata
